[Refactor] 프로젝트 구조 수정
