+++ src/main/java/verbly/spring/domain/payment/service/PayPalScheduler.java
@@ -4,9 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import verbly.spring.domain.payment.client.PaypalClient;
-import verbly.spring.domain.payment.dto.PaypalDTO;
 import verbly.spring.domain.payment.entity.Subscription;
 import verbly.spring.domain.payment.enums.PaymentProvider;
 import verbly.spring.domain.payment.enums.SubscriptionStatus;
