+++ src/main/java/verbly/spring/domain/payment/service/PayPalService.java
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.util.UriComponentsBuilder;
-import verbly.spring.domain.payment.client.PaypalClient;
-import verbly.spring.domain.payment.dto.PaypalDTO;
+import verbly.spring.global.infrastructure.paypal.PaypalClient;
+import verbly.spring.global.infrastructure.paypal.dto.PaypalDTO;
 import verbly.spring.domain.payment.entity.Subscription;
 import verbly.spring.domain.payment.entity.SubscriptionPlan;
 import verbly.spring.domain.payment.enums.BillingCycle;
